Add support for rate-based simulations
Rate-based simulations run by creating a number of requests per second,
regardless of the state of existing requests, rather than maintaining a
number of concurrent requests.

Currently, it seems to be possible for rate based simulations to 'lock
up' if too many requests block for too long. I _think_ what is happening
is that all the threads created by `pipeline/prun` are blocked waiting
for a response, so no new requests are made until they return. I'm not
100% sure why this is happening. We pass in users equal to rate *
timeout, so by the time we run out of 'free' threads the oldest one
should have timed out and be ready to issue a new request. We are using
`core.async`, but the lib still expects each step to synchronously
return a boolean before either deciding to run the next step or a new
instance of the scenario.

Testing against https://www.google.com/ at 100 reqs/sec for 30s showed
some requests taking 11+ secs, even with a timeout of 5s. Is there some
additional timeout we're hitting that isn't covered by `:timeout` in
http-kit and defaults to much higher?