Improve Gatling Highcharts reporting
This uses the modifications to the Gatling Highcharts reporter
(https://github.com/mhjort/clojider-gatling-highcharts-reporter/pull/2)
to record a breakdown of the types of failure that occur. User-defined
request functions simply need to throw the exception for this to be
triggered.

Given that some libraries (e.g. `httpkit`) return a clojure
ExceptionInfo in a failed response it is also useful to check the
response for a directly returned exception object, rather than requiring
it to be thrown.